Refactor to have make_sgf take a string
This should put me in a good position to write a bunch of end-to-end
tests.

Obviously these tests will be brittle to any changes in the output but
once whenever I make a change that alters the output I should be able to
generate all the expected SVGs and then visually check them quickly to
make sure they look right before updating the test output.