Switch toml styles and make tests use args
This commit switches to styles read from TOML, and has the tests take
command line arguments as inputs instead of a serialized options struct.

This lets me test the end-to-end behavior better, and also avoids the
issue of having to specify the styles as a struct in the tests when all
I really care about is the style name.

This commit breaks the fancy style since I don't yet have support for
writing defs. It turns out that the SVG library I'm using doesn't really
support reading and then writing. See
https://github.com/bodoni/svg/issues/41. To get defs support I'm either
going to have to add some sort of element builder layer on top of the
event based parser the library exposes, or, more likely switch libraries
to a different XML library.

Given that my building logic mostly operates on a DOM level, I'm leaning
towards a library that supports a DOM model (not a pure event model).
https://crates.io/crates/RustyXML seems like a good choice.