Refactor goban
This commit doesn't change any behavior, but it removes all direct
access to `Goban` fields, and simplifies a bunch of the associated
logic.

This should put me in a better position to work on fixing `--kifu` mode
when using `--first-move-number` (issue #18) which was feeling very
complicated with the old logic.