Fix kifu repeated move formatting when missing board labels
The way options were being parsed it was possible to have board labels
disabled, but still have `label_sides` set with margins calculated for
the non-rendering sides. I've removed the `draw_board_labels` option in
favor of just looking for an empty `label_sides` config, and the
argument parsing handles that.