Bump some dependencies
Clippy was complaining that tiny-xml (required by minidom) had code that
was going to break. Updating minidom to the latest version fixed the
issue, though it also made small changes to all the svg output.

I've visually verified that the new output is still correct, and updated
the tests to match.

WHile I was at it, I updated the other misc dependencies, but skipped
`resvg` and related packages since those involve breaking changes I'm
not looking to fix right this moment.