build CI -> create release CI
I basically never want to just build, so I may as well create the
release in the process!

This will just create a draft release, so I still have to go in and
write the changelog and do the `cargo publish`, but it should save most
of the boring work.