Build man page automatically
Instead of commiting the built man page to the repo, build it
automatically as part of the release process when making a .deb.

`scripts/build-deb.sh` does the whole job of generating the man page and
building a .deb file. You can still generate the man page can separately
with `scripts/generate-man.sh`.

Note that `cargo deb` probably shouldn't be used directly anymore since
if there's an out-of-date man page sitting at `man/sgf-render.1.gz`
`cargo deb` will happily generate a package with the old man page. As
far as I can tell there's no way to prevent that while maintaining the
ability to run `generate-man.sh` separately, but it's not the end of the
world.